Move SingleChipLayouter into a submodule
It is a single-pass layouter, but we won't rename it for now.
